backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# HTTP + scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

# AI / embeddings
google-generativeai==0.7.2

# RAG orchestration
langchain==0.2.5
langchain-community==0.2.5

# Database
supabase==2.4.6

# Utilities
python-dotenv==1.0.1
tiktoken==0.7.0
tqdm==4.66.4
pydantic==2.7.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
respx==0.21.1
pytest-mock==3.14.0